fixed author verification

+++ src/comments/comments.service.ts
@@ -30,7 +30,6 @@ export class CommentsService {
         //let comment = await this.commentsRepository.findOne({where: {id: dto.commentId}});
         const user = this.jwtService.verify(refreshToken, {secret: "FFFGKJKFWMV"});
         let comment = await this.commentsRepository.findOne({where: {id: dto.commentId}});
-        console.log(user.displayName);
         let isAuthor = await this.commentsRepository.findOne({where: {id: dto.commentId, displayName: user.displayName}})
         if (!comment) {
             throw new HttpException('Комментарий не найден', HttpStatus.NOT_FOUND);
@@ -41,11 +40,16 @@ export class CommentsService {
         await this.commentsRepository.update({comment: dto.comment}, {where: {id: dto.commentId}});
     }
 
-    async deleteComment(commentId: number) : Promise<void>  {
+    async deleteComment(commentId: number, refreshToken: string) : Promise<void>  {
+        const user = this.jwtService.verify(refreshToken, {secret: "FFFGKJKFWMV"});
         let comment = await this.commentsRepository.findOne({where: {id: commentId}});
+        let isAuthor = await this.commentsRepository.findOne({where: {id: commentId, displayName: user.displayName}});
         if(!comment) {
             throw new HttpException('Комментарий не найден', HttpStatus.NOT_FOUND);
         } 
+        if (!isAuthor) {
+            throw new HttpException('Нет доступа', HttpStatus.FORBIDDEN);
+        }
         await this.commentsRepository.destroy({where: {id: commentId}});
     }
